+++ src/shapeSearch/gpu/copyDescriptorsToHost.h
@@ -3,4 +3,4 @@
 #include <shapeSearch/common/types/array.h>
 #include <shapeSearch/libraryBuildSettings.h>
 
-array<newSpinImagePixelType> copyQSIDescriptorsToHost(array<newSpinImagePixelType> device_descriptors, unsigned int imageCount);
+array<newSpinImagePixelType> copyQSIDescriptorsToHost(array<newSpinImagePixelType> device_descriptors, size_t imageCount);
